more consensus type consolidation
